Refactor EditBookCommand and EditBookCommandParser into the book packages.
Currently the classes which concern Edit Command is not segregated into another package.

This makes the code base less organised.

 Let's
*Refactor EditBookCommand and EditBookCommandParser into book packages
